[iss-202] extension support as Bottle plugin
