+++ tests/ext/bottle/views/index.tpl
@@ -1,7 +1,2 @@
-%if name == 'World':
-    <h1>Hello {{name}}!</h1>
-    <p>This is a test.</p>
-%else:
-    <h1>Hello {{name.title()}}!</h1>
-    <p>How are you?</p>
-%end
+<h1>Hello {{name.title()}}!</h1>
+<p>How are you?</p>
