+++ tox.ini
@@ -13,7 +13,7 @@ deps =
     coverage==4.5.4
     codecov
     requests
-    bottle >= 0.12.13
+    bottle >= 0.10
     flask >= 0.10
     sqlalchemy
     Flask-SQLAlchemy
